Merged PR 154400: Change block  comment to line comments
Thank you for your contribution to the iTwinUi project! Please make sure to complete all of the items below before submitting the PR.

## Checklist

- :x: Add visual tests for your component
- ☑ Verify that all existing tests pass
- :x: Add an entry to the changelog for any new components and changes
- :x: Add screenshots of the key elements of the component below
- ☑ Add any additional comments below describing the features you've implemented
- ☑ Link work item

## Screenshots

## Additional Comments
Only changed the block comment for the "themed" function so processors automatically skip that comment from output.

Changed block  comment to line comments, the only impact I can think of is if someone managed to have working intellisense, that might break it, but given that this is the only function with such comment type, I suppose it will not really break anything.

Related work items: #586147