refactor: Use inline svgs for checkbox and radio (#475)
Simplified the code for checkbox and radio so that it doesn't rely on 0x0 checkbox and can be used without a label. The markup no longer needs svgs as they are inlined in the CSS. No visual changes in modern browsers. IE gets to show the default checkbox.

Class name changes:

iui-checkbox -> iui-checkbox-wrapper
iui-radio -> iui-radio-wrapper
iui-label -> iui-checkbox-label and iui-radio-label
New: input.iui-checkbox and input.iui-radio
Removed: iui-checkbox-checkmark and iui-radio-dot