fix(Carousel): Use event handler instead of useEffect for scroll (iTwin/iTwinUI-react#683)
useEffect was causing problems on React 18, so this change reworks the scrolling mechanism to explicitly call scrollTo on every single manual event (clicks, keyboard) as well as when the currentIndex prop changes. It also sets isManuallyUpdating to true. And then we listen to the 'scroll' event and reset isManuallyUpdating to false after the very last scroll event.

The intersection observer is used for syncing the active dot with scroll events that are triggered natively (e.g. dragging on mobile). We check for isManuallyUpdating here to avoid messing with manual scrollTo calls.