chore: Update test images and fix bad rebase
While rebasing main branch onto v1, it was hard to fix merge conflicts so I ended up just taking images from the main branch instead of running tests on each conflict and approving combined images. This commit updates all the images for all the changes in the v1 branch so far together.

Also there were a couple bad merges in the rebase process so those are corrected as well.