feat(ThemeProvider): Add `applyBackground` to `themeOptions` (iTwin/iTwinUI-react#974)
Added an optional boolean applyBackground to themeOptions. When set to true, it will add the iui-root-background class along with iui-root. When false, it will not. The default behavior (when user doesn't pass anything) is to check if theme was already set either through context or on <body> and add this class if this is the first time setting theme. This makes it easier for the user to just drop in a ThemeProvider at the top of their app.