chore: Update deps (#857)
Updated deps in preparation of monorepo merge. It's important not to have conflicts especially when it comes to prettier and husky. Husky hook format changed (moved from package.json to .husky folder) and prettier resulted in some formatting changes in vite.config.ts and badge.js. Other than that, everything seems to work the same.