Update PR template (#977)
Instead of a boring checklist which gets ignored in all PRs, I've separated the PR template into three sections (much like this very PR description) and added somewhat helpful comments to each. If one of the sections does not apply to a PR, then we can write "N/A". Also explicitly called out that the template can be ignored for purely docs site changes, so as to not confuse non-technical contributors.