replace sandpack with static examples (#1088)
Sandpack comes with a huge bundle and takes forever to load, so I'm replacing it with statically rendered, hydrated examples (they load instantly and are also interactive). For editing, they can be opened in codesandbox externally.

In the process, I also improved a lot of the example code for readability/usability.

The approach for including code examples has now changed. Examples will first need to be exported in src/examples/index.tsx and then used in mdx files manually.