refactor `ProgressRadial` to use pseudo-elements (#1356)
Instead of using an svg, the radial progress indicator will now use pseudo-elements with a combination of conic and linear gradients, greatly simplifying both the markup and styles.

The size/status is set using data attributes. For the determinate variant, the percentage value will be set using a CSS variable `--iui-progress-percentage`.