remove `overflow` and inline `transform` from Dialog (#1795)
(backported to v2)

This fixes blurring and avoids issues with dialog becoming off-center when the window is resized.

The overflow is already handled by Dialog.Content, which is required in v3 but optional in v2. For this reason, I've kept some of the checks around (even though they were removed in #1795).